[NAYB-38] refactor : fixture 파일 이동
