+++ src/test/java/com/prgrms/nabmart/domain/event/service/EventItemServiceTest.java
@@ -17,8 +17,8 @@ import com.prgrms.nabmart.domain.item.domain.Item;
 import com.prgrms.nabmart.domain.item.exception.NotFoundItemException;
 import com.prgrms.nabmart.domain.item.repository.ItemRepository;
 import com.prgrms.nabmart.global.fixture.CategoryFixture;
-import com.prgrms.nabmart.global.fixture.EventFixture;
-import com.prgrms.nabmart.global.fixture.EventItemFixture;
+import com.prgrms.nabmart.domain.event.support.EventFixture;
+import com.prgrms.nabmart.domain.event.support.EventItemFixture;
 import com.prgrms.nabmart.global.fixture.ItemFixture;
 import java.util.Arrays;
 import java.util.Optional;
